move arg parsing to separate file
