+++ src/main.hpp
@@ -1,14 +1,16 @@
 #ifndef CCC_MAIN_HPP_INCLUDED
 #define CCC_MAIN_HPP_INCLUDED
 
-struct cmd_line_args
-{
-	int printflag = 0;
-	int lexflag = 0;
-	int printir = 0;
-	int optlevel = 1;
-	int keep_pp = 0;
-	char* filename = NULL;
-};
+// struct cmd_line_args
+// {
+// 	int printflag = 0;
+// 	int lexflag = 0;
+// 	int printir = 0;
+// 	int optlevel = 1;
+// 	int keep_pp = 0;
+// 	char* filename = NULL;
+// };
+
+#define noop 0
 
 #endif // CCC_MAIN_HPP_INCLUDED
